fix: entity 칼럼 수정 not null 등 엔티티 칼럼 전체 수정하였습니다.

+++ src/main/java/com/example/team16project/dto/article/ArticleDto.java
@@ -11,7 +11,7 @@ import java.time.LocalDateTime;
 @Data
 @Builder
 public class ArticleDto {
-    Integer articleId;
+    Long articleId;
     User user;
     String title;
     String contents;
